feat: move core-wrappers to core-components
Because the distinction between them is not accurate.

Both have components that import other components.
