+++ apps/ui-patterns/src/app/UIPatternsComplexWizard/index.ts
@@ -34,7 +34,7 @@ export type InitialValueGenerator = (hookValues: {
   extra: UIWizardExtra;
 }) => Partial<UIWizardSchema>;
 
-// An adapter type for mapping additional values to @tacc/core-wrappers/Wizard
+// An adapter type for mapping additional values to @tacc/core-components/Wizard
 export type UIWizardStep = {
   id: string;
   name: string;
