+++ apps/ui-patterns/src/app/UIPatternsComplexWizard/steps/StepTwo.tsx
@@ -9,7 +9,7 @@ import { Button } from '@tacc/core-components';
 import fieldArrayStyles from './FieldArray.module.css';
 import { Collapse } from '@tacc/core-components';
 import { FieldArray, useFormikContext, FieldArrayRenderProps } from 'formik';
-import { FormikInput, FormikCheck } from '@tacc/core-wrappers';
+import { FormikInput, FormikCheck } from '@tacc/core-components';
 import {
   UIWizardStep,
   UIWizardComplexField,
