+++ apps/ui-patterns/src/app/UIPatternsWizard/steps/LastStep.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useState, useMemo } from 'react';
 import * as Yup from 'yup';
 import { UIWizardSchema, useWizardValues } from '..';
 import { Button } from '@tacc/core-components';
-import { SubmitWrapper } from '@tacc/core-wrappers';
-import { WizardStep } from '@tacc/core-wrappers';
+import { SubmitWrapper } from '@tacc/core-components';
+import { WizardStep } from '@tacc/core-components';
 import { validationSchema as stepOneValidationSchema } from './StepOne';
 import { validationSchema as stepTwoValidationSchema } from './StepTwo';
 import { validationSchema as stepThreeValidationSchema } from './StepThree';
