+++ libs/core-components/src/index.ts
@@ -1,3 +1,5 @@
+import { WizardStep as WizardStepType } from './lib/Wizard';
+
 export { default as Button } from './lib/Button';
 export { default as Icon } from './lib/Icon';
 export { default as Section } from './lib/Section';
@@ -18,5 +20,19 @@ export { default as HistoryBadge } from './lib/HistoryBadge';
 export { default as Collapse } from './lib/Collapse';
 export { default as TextCopyField } from './lib/TextCopyField';
 export * from './lib/Form';
+
+export { Navbar, NavItem, QueryNavItem, AnchorNavItem } from './lib/Navbar';
+export { default as QueryWrapper } from './lib/QueryWrapper';
+export { default as SubmitWrapper } from './lib/SubmitWrapper';
+export { default as Wizard, useWizard, WizardNavigation } from './lib/Wizard';
+export type WizardStep<T> = WizardStepType<T>;
+export {
+  FormikFieldWrapper,
+  FormikInput,
+  FormikSelect,
+  FormikCheck,
+  FormikTextarea,
+  FormikFileInput,
+} from './lib/FormikFieldWrapper';
 
 export { default as withBuilder } from './utils/withBuilder';
