+++ libs/core-components/src/lib/Navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import { Icon } from '@tacc/core-components';
-import styles from './NavbarWrapper.module.css';
+import styles from './Navbar.module.css';
 
 type sizes = 'xs' | 'sm' | 'md' | 'lg';
