+++ libs/core-components/src/lib/Navbar/index.ts
@@ -1,0 +1,2 @@
+export { default as Navbar } from './Navbar';
+export { NavItem, QueryNavItem, AnchorNavItem } from './Navbar';
