--- libs/core-wrappers/README.md
@@ -1,7 +1,0 @@
-# core-wrappers
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test core-wrappers` to execute the unit tests via [Vitest](https://vitest.dev/).
