--- libs/core-wrappers/src/index.ts
@@ -1,19 +1,0 @@
-import { WizardStep as WizardStepType } from './lib/Wizard';
-export {
-  Navbar,
-  NavItem,
-  QueryNavItem,
-  AnchorNavItem,
-} from './lib/NavbarWrapper';
-export { default as QueryWrapper } from './lib/QueryWrapper';
-export { default as SubmitWrapper } from './lib/SubmitWrapper';
-export { default as Wizard, useWizard, WizardNavigation } from './lib/Wizard';
-export type WizardStep<T> = WizardStepType<T>;
-export {
-  FormikFieldWrapper,
-  FormikInput,
-  FormikSelect,
-  FormikCheck,
-  FormikTextarea,
-  FormikFileInput,
-} from './lib/FormikFieldWrapper';
