--- libs/core-wrappers/src/lib/NavbarWrapper/index.ts
@@ -1,2 +1,0 @@
-export { default as Navbar } from './NavbarWrapper';
-export { NavItem, QueryNavItem, AnchorNavItem } from './NavbarWrapper';
