+++ libs/tup-components/src/tickets/TicketDetailModal/TicketReplyForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Formik, Form, FormikHelpers, Field } from 'formik';
-import { FormikFileInput, FormikTextarea } from '@tacc/core-wrappers';
+import { FormikFileInput, FormikTextarea } from '@tacc/core-components';
 import { FormGroup } from 'reactstrap';
 import { Button, InlineMessage } from '@tacc/core-components';
 import { useTicketReply } from '@tacc/tup-hooks';
